+++ src/layouts/Index.vue
@@ -83,7 +83,7 @@ export default {
       toolboxMode: undefined,
       connectFlag: false,
       isTabsVisible: true,
-      entityByGroup: ['platform', 'webhooks', 'realms', 'grants', 'channels', 'devices', 'groups', 'streams', 'calcs', 'intervals', 'plugins', 'hexViewer', 'trafficViewer', 'deviceTraffic', 'streamTraffic', 'webhookTraffic', 'modems', 'containers', 'cdns', 'mqtt', 'mqttClient'],
+      entityByGroup: ['platform', 'webhooks', 'realms', 'grants', 'channels', 'devices', 'groups', 'streams', 'calcs', 'intervals', 'plugins', 'hexViewer', 'trafficViewer', 'deviceTraffic', 'streamTraffic', 'webhookTraffic', 'pluginTraffic', 'modems', 'containers', 'cdns', 'mqtt', 'mqttClient'],
       isNeedSelect: true,
       entityInited: false,
       isInit: Vue.connector.socket.connected()
@@ -205,6 +205,7 @@ export default {
               result.push('trafficViewer')
               result.push('deviceTraffic')
               result.push('streamTraffic')
+              result.push('pluginTraffic')
               break
             }
             case 'storage': {